tmp/stt/pyproject.toml
[project]
name = "stt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper>=20240930",
    "pyaudio>=0.2.14",
    "pyttsx3>=2.98",
    "ruff>=0.11.8",
]
